\name{bivariate-descriptive-statistics-with-R}
\alias{bivariate-descriptive-statistics-with-R}
\title{
Tutorial one applying R commands for ivariate descriptive statistics
}
\description{
Learn how to apply R to do descriptive statistics for one variable.
}
\usage{
run_tutorial("bivariate-descriptive-statistics-with-R", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("bivariate-descriptive-statistics-with-R", package = "FaRTutorials")
}
