\name{data-basic}
\alias{data-basic}
\title{Learnr tutorial about the base data types in R.}
\description{
In this tutorial, you will learn how to use R to inspect the contents of a data frame or tibble. Data frames and tibbles are R's structures for storing tabular data; if you inherit a tabular dataset in R, it will almost certainly come as one of these structures.

Here, you will learn how to do three things with data frames and tibbles:

1. Look at the contents of a data frame or tibble

2. Open a help page that describes a data frame or tibble

3. Identify the variables and their types in a tibble

You will also meet the `mpg` and `flights` datasets. These datasets appear frequently in R examples.
}
\usage{
run_tutorial("data-basic", package = "FaRTutorials")
}
\author{
Part of learnr package: https://rstudio.github.io/learnr/
}
\examples{
run_tutorial("data-basics", package = "FaRTutorials")
}
\references{
The readings in this tutorial follow [_R for Data Science_](http://r4ds.had.co.nz/), sections 3.2 and 5.1.
source: https://github.com/rstudio/learnr/tree/master/inst/tutorials/
}
