\name{data-filter}
\alias{data-filter}
\title{
Learnr tutorial about filtering observations
}
\description{
In this tutorial, you will learn how to:

* use `filter()` to extract observations from a data frame or tibble
* write logical tests in R
* combine logical tests with Boolean operators
* handle missing values within logical tests
}
\usage{
run_tutorial("data-filter", package = "FaRTutorials")
}
\author{
The readings in this tutorial follow [_R for Data Science_](http://r4ds.had.co.nz/), section 5.2.
source: Part of learnr package: https://rstudio.github.io/learnr/
}
\examples{
run_tutorial("data-filter", package = "FaRTutorials")
}
\references{
https://github.com/rstudio/learnr/tree/master/inst/tutorials/
}
