\name{data-summarise}
\alias{data-summarise}
\title{
Learnr tutorial about  summarizing the columns available in an R data frame
}
\description{
In this tutorial, you will learn how to summarise a table of data, including:

* How to summarise tables with `summarise()`
* How to recognize the most useful families of functions to combine with `summarise()`
* How to combine multiple dplyr operations with the pipe, `%>%`
* How to compute counts of observations with `n()`
* How to summarise groups of observations with `group_by()` and `summarise()`
}
\usage{
run_tutorial("data-summarise", package = "FaRTutorials")
}
\author{
The readings in this tutorial follow [_R for Data Science_](http://r4ds.had.co.nz/), section 5.6.
source: Part of learnr package: https://rstudio.github.io/learnr/
}
\examples{
run_tutorial("data-summarise", package = "FaRTutorials")
}
\references{
https://github.com/rstudio/learnr/tree/master/inst/tutorials/
}
