\name{data-mutate}
\alias{data-mutate}
\title{
Learnr tutorial about deriving new variables from a data frame
}
\description{
In this tutorial, you will learn how to derive new variables from a data frame, including:

* How to create new variables with `mutate()`
* How to recognize the most useful families of functions to use with `mutate()`
}
\usage{
run_tutorial("data-mutate", package = "FaRTutorials")
}
\author{
Part of learnr package: https://rstudio.github.io/learnr/
}
\examples{
run_tutorial("data-mutate", package = "FaRTutorials")
}
\references{
The readings in this tutorial follow [_R for Data Science_](http://r4ds.had.co.nz/), section 5.5.
Source: https://github.com/rstudio/learnr/tree/master/inst/tutorials/
}
