\name{normal_distr}
\alias{normal_distr}
\title{
Tutorial one normal distribution and their properties
}
\description{
Learnr tutorial for computing probaility and quantile of normally distributed random variables
}
\usage{
run_tutorial("normal_distr", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("normal_distr", package = "FaRTutorials")
}
