\name{R-programming-language}
\alias{R-programming-language}
\title{
Tutorial one RStudio using a simple script and on the use of the console as a pocket calculator
}
\description{
Learn how to use RSTudio and arithmetic operations.
}
\usage{
run_tutorial("R-programming-language", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("R-programming-language", package = "FaRTutorials")
}
