\name{R_first_steps}
\alias{R_first_steps}
\title{
Tutorial one RStudio using a simple script and on the use of the console as a pocket calculator
}
\description{
Learn how to use RSTudio and arithmetic operations.
}
\usage{
run_tutorial("R_first_steps", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("R_first_steps", package = "FaRTutorials")
}
