\name{tests_conf_intervals}
\alias{tests_conf_intervals}
\title{
Tutorial one sample tests and confidence intervals
}
\description{
Learnr tutorial for computing one sample tests and confidence in the normal
and in bernoulli model}
\usage{
run_tutorial("tests_conf_intervals", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("tests_conf_intervals", package = "FaRTutorials")
}
