\name{tidy_data}
\alias{tidy_data}
\title{
Tutorial one example to tidy data applying tidyverse functions
}
\description{
Learn how to use tidyverse function to tidy messy data
}
\usage{
run_tutorial("tidy_data", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("tidy_data", package = "FaRTutorials")
}
