\name{tutorial_case_study}
\alias{tutorial_case_study}
\title{
Tutorial one data cleaning and tidying data by a case study
}
\description{
Learnr tutorial for applying the commands gather(), separate(), mutate(), right_join()

Clean and tidy a subset of data from the World Health Organization Global Tuberculosis Report,
and accompanying global populations.
}
\usage{
run_tutorial("tutorial_case_study", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("tutorial_case_study", package = "FaRTutorials")
}
