\name{univariate-descriptive-statistics-with-R}
\alias{univariate-descriptive-statistics-with-R}
\title{
Tutorial one applying R commands for univariate descriptive statistics
}
\description{
Learn how to apply R to do descriptive statistics for one variable.
}
\usage{
run_tutorial("univariate-descriptive-statistics-with-R", package = "FaRTutorials")
}
\author{
Dr. Egbert Falkenberg
}
\examples{
run_tutorial("univariate-descriptive-statistics-with-R", package = "FaRTutorials")
}
